myapp/flask_app/requirements.txt
Flask==2.0.3
redis==4.1.0
psycopg2-binary==2.9.1
